cleaned up fx krystal for mae....

+++ cogs/MaePings.py
@@ -32,14 +32,9 @@ class MaePings(commands.Cog):
         
     @commands.command()
     async def krystal(self, ctx):
-        if ctx.guild.id == luminary:
-            if ctx.channel.id == kbotcom:
-                await ctx.send(f'<@{mae}>, <@!{ctx.author.id}> is talking about Krystal :gem:')
-                await ctx.send(random.choice(self.bot.krystal_gif))
-                await ctx.message.delete()
-            else:
-                await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
-                await ctx.message.delete()
+        if ctx.guild.id == luminary and ctx.channel.id != kbotcom:
+            await ctx.send(content=f'Wrong channel <@!{ctx.author.id}>! Go to <#764610881513324574>', delete_after=2)
+            await ctx.message.delete()
         else:
             await ctx.send(f'<@!{ctx.author.id}> is talking about Krystal :gem:')
             await ctx.send(random.choice(self.bot.krystal_gif))
